requirements.txt
mysql-connector-python==8.2.0
psycopg2-binary==2.9.7
pandas==2.1.3
numpy==1.24.3